Refactor: make get_unet and load_weights importable without running tests
